@@ -1,7 +1,9 @@
 package com.udacity.pricing.domain.price;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+
 import java.math.BigDecimal;
 
 /**
